reorganize code to reduce number of modules
