+++ magma/lib/magma/query/predicate.rb
@@ -407,7 +407,6 @@ require_relative 'predicate/subquery_operator'
 require_relative 'predicate/subquery_utils'
 require_relative 'predicate/subselect/boolean_subselect'
 require_relative 'predicate/subselect/child_model_subselect'
-require_relative 'predicate/subselect/column_subselect'
 require_relative 'predicate/subselect/date_time_subselect'
 require_relative 'predicate/subselect/file_collection_subselect'
 require_relative 'predicate/subselect/file_subselect'
