+++ magma/lib/magma/query/predicate/file.rb
@@ -1,10 +1,15 @@
 require_relative '../md5_set'
 require_relative '../updated_at_set'
-require_relative 'with_file_predicate_methods'
 
 class Magma
   class FilePredicate < Magma::ColumnPredicate
-    include WithFilePredicateMethods
+    attr_reader :requested_file_paths
+
+    def initialize question, model, alias_name, attribute, *query_args
+      super
+      @md5_set = Md5Set.new(@question.user, @model)
+      @updated_at_set = UpdatedAtSet.new(@question.user, @model)
+    end
 
     verb '::url' do
       child String
