+++ magma/lib/magma/query/predicate/subselect/boolean_subselect.rb
@@ -1,7 +1,10 @@
-require_relative 'column_subselect'
+require_relative '../boolean'
+require_relative 'with_subselect_override_methods'
 
 class Magma
-  class BooleanSubselectPredicate < Magma::ColumnSubselectPredicate
+  class BooleanSubselectPredicate < Magma::BooleanPredicate
+    include WithSubselectOverrideMethods
+
     def self.verbs
       Magma::BooleanPredicate.verbs
     end
