+++ magma/lib/magma/query/predicate/subselect/file_subselect.rb
@@ -1,9 +1,9 @@
-require_relative 'column_subselect'
-require_relative '../with_file_predicate_methods'
+require_relative 'with_subselect_override_methods'
+require_relative '../file'
 
 class Magma
-  class FileSubselectPredicate < Magma::ColumnSubselectPredicate
-    include WithFilePredicateMethods
+  class FileSubselectPredicate < Magma::FilePredicate
+    include WithSubselectOverrideMethods
 
     def self.verbs
       Magma::FilePredicate.verbs
