+++ magma/lib/magma/query/predicate/subselect/match_subselect.rb
@@ -1,7 +1,10 @@
-require_relative 'column_subselect'
+require_relative '../match'
+require_relative 'with_subselect_override_methods'
 
 class Magma
-  class MatchSubselectPredicate < Magma::ColumnSubselectPredicate
+  class MatchSubselectPredicate < Magma::MatchPredicate
+    include WithSubselectOverrideMethods
+
     def self.verbs
       Magma::MatchPredicate.verbs
     end
