+++ magma/lib/magma/query/predicate/subselect/number_subselect.rb
@@ -1,7 +1,10 @@
-require_relative 'column_subselect'
+require_relative 'with_subselect_override_methods'
+require_relative '../number'
 
 class Magma
-  class NumberSubselectPredicate < Magma::ColumnSubselectPredicate
+  class NumberSubselectPredicate < Magma::NumberPredicate
+    include WithSubselectOverrideMethods
+
     def self.verbs
       Magma::NumberPredicate.verbs
     end
