+++ magma/lib/magma/query/predicate/subselect/string_subselect.rb
@@ -1,7 +1,10 @@
-require_relative 'column_subselect'
+require_relative 'with_subselect_override_methods'
+require_relative '../string'
 
 class Magma
-  class StringSubselectPredicate < Magma::ColumnSubselectPredicate
+  class StringSubselectPredicate < Magma::StringPredicate
+    include WithSubselectOverrideMethods
+
     def self.verbs
       Magma::StringPredicate.verbs
     end
