--- magma/lib/magma/query/predicate/with_date_time_predicate_methods.rb
@@ -1,7 +1,0 @@
-class Magma
-  module WithDateTimePredicateMethods
-    def extract table, identity, is_all
-      Magma::Answer.new(table.first[column_name]&.iso8601)
-    end
-  end
-end
